+++ divisao/src/divi_07_2.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int divi (int a, int b){
-float res;
+int res;
 res = a/b;
 return (res);
 }
