+++ divisao/src/divi_09_2.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-float divi (float div1,float dividendo){
+int divi (int div1,int dividendo){
 div1 =  div1/dividendo;
 return (div1);
 }
